finished baseline project

+++ src/com/skilldistillery/foodtruck/FoodTruck.java
@@ -7,7 +7,7 @@ public class FoodTruck {
 
 	private String name;
 	private String foodType;
-	private double rating;
+	private int rating;
 	private int truckID;
 	private static int numID = 1;
 
@@ -20,7 +20,7 @@ public class FoodTruck {
 		numID++;
 	}
 
-	public FoodTruck(String name, String foodType, double rating) {
+	public FoodTruck(String name, String foodType, int rating) {
 		this.name = name;
 		this.foodType = foodType;
 		this.rating = rating;
@@ -46,11 +46,11 @@ public class FoodTruck {
 		this.foodType = foodType;
 	}
 
-	public double getRating() {
+	public int getRating() {
 		return rating;
 	}
 
-	public void setRating(double rating) {
+	public void setRating(int rating) {
 		this.rating = rating;
 	}
